+++ activesupport/lib/active_support/core_ext/date/calculations.rb
@@ -142,4 +142,6 @@ class Date
   end
   alias_method :compare_without_coercion, :<=>
   alias_method :<=>, :compare_with_coercion
+  alias_method :before?, :<
+  alias_method :after?, :>
 end