+++ activesupport/CHANGELOG.md
@@ -6,6 +6,11 @@
 
     *Ashe Connor*, *Aaron Patterson*
 
+*   Add `before?` and `after?` methods to `Date`, `DateTime`,
+    `Time`, and `TimeWithZone`.
+
+    *Nick Holden*
+
 *   Add `:private` option to ActiveSupport's `Module#delegate`
     in order to delegate methods as private: